% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_log_entry}
\alias{build_log_entry}
\title{Build a single processing log entry}
\usage{
build_log_entry(step, n_cells, n_genes, params)
}
\arguments{
\item{step}{Name of the processing step}

\item{n_cells}{Number of cells in the object at this step}

\item{n_genes}{Number of genes in the object at this step}

\item{params}{Named list of parameters used in this step}
}
\value{
A named list representing one log entry
}
\description{
Build a single processing log entry
}
\keyword{internal}
