% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_passport}
\alias{build_passport}
\title{Build a passport list from components}
\usage{
build_passport(
  object_id,
  rds_self,
  animal_id,
  species,
  sex,
  age,
  condition,
  tissue,
  project,
  researcher,
  date,
  notes,
  parent_id,
  rds_parent,
  lineage,
  children,
  rds_children,
  custom_keys,
  custom_vals
)
}
\arguments{
\item{object_id}{Name/ID of this Seurat object}

\item{rds_self}{RDS registry number of this object}

\item{animal_id}{Individual animal identifier}

\item{species}{Species name}

\item{sex}{Sex of the animal}

\item{age}{Age of the animal}

\item{condition}{Experimental condition}

\item{tissue}{Tissue of origin}

\item{project}{Project name}

\item{researcher}{Name of the researcher}

\item{date}{Date of the experiment}

\item{notes}{Free-text notes}

\item{parent_id}{Object ID of the direct parent}

\item{rds_parent}{RDS number of the parent object}

\item{lineage}{Full ancestry chain as character vector}

\item{children}{Object IDs of children}

\item{rds_children}{RDS numbers of children objects}

\item{custom_keys}{Names of custom fields}

\item{custom_vals}{Values of custom fields}
}
\value{
A named list representing the passport
}
\description{
Build a passport list from components
}
\keyword{internal}
