% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate_passport}
\alias{validate_passport}
\title{Check whether a passport list is valid}
\usage{
validate_passport(passport)
}
\arguments{
\item{passport}{A list (the \code{@misc$passport} slot of a Seurat object).}
}
\value{
TRUE if the passport contains all required fields, FALSE otherwise.
}
\description{
Check whether a passport list is valid
}
\keyword{internal}
